requirements.txt
ultralytics
gTTS
pygame
opencv-python
torch
requests
numpy
Pillow
transformers
playsound
deep-translator
streamlit
watchdog
